Some changes to make the battle AI smarter
- the AI will now consider attacking multiple units
- the preferred strategy now is to minimize collateral damage rather than to maximize damage to enemy units alone
- attacks that block enemy shooters will be prioritized over other attacks in cases when shooters have weaker melee attacks

+++ AI/BattleAI/PotentialTargets.h
@@ -20,5 +20,5 @@ public:
 	PotentialTargets(const battle::Unit * attacker, const HypotheticBattle * state);
 
 	AttackPossibility bestAction() const;
-	int bestActionValue() const;
+	int64_t bestActionValue() const;
 };
